Add prompt argument to Open Connection calls
Add prompt argument to SSHLibrary's Open Connection calls. This makes
the SSH connection succeed even if it takes more than 0.5 seconds for
the actual prompt to appear after SSH connection has been established,
and it will also make test faster if it appears earlier. It will wait
for prompt to appear for 30 seconds.

Signed-off-by: Mika Tammi <mika.tammi@unikie.com>