PoC gui test automation
Create gui test suite with common setup. Create
generic keywords for launching and closing any app
with gui test automation.

Create example test cases for chromium and firefox.
Added test case for logout and login, and test for
booting to login screen.

Test case for GUI reboot is added. This test case
verifies also that the device boots to login screen,
not directly to desktop.

Login is included in the suite setup because nearly
all gui tests depend on login. However, login is also
added as a separate test case (after reboot test case).

Log out is added as a separate test case. It should
be taken care that this test will be the last one
in the suite.

Gui tests are expected to be run from logged out
state but the login keyword is designed so that it
does not fail even when run in already logged in state.

Graphical details and icons used in image recognition
are copied from run-time ghaf gui-vm.

Added new ssh keyword 'Connect to VM if not already
connected' to speed up testing by reducing unnecessary
re-connect cycles.

Fixed also a bug in iperf test report message showing
only tx results.

Signed-off-by: Samuli Leivo <samuli.leivo@unikie.com>