Add iterations to Connect keyword
After ethernet interface was passed through to net-vm on
lenovo-x1 iterative check of login to net-vm was first
added to boot tests of lenovo-x1. Fails after gui reboot
tests showed that connection iterations need be added
also after gui reboot, namely every time the device boots
or reboots.

Decided to put the default number of iterations in Connect
keyword to 1 (as before) and give more iterations only
when it is used after boot or reboot.

Signed-off-by: Samuli Leivo <samuli.leivo@unikie.com>