Prevent failed iteration to fail the keyword
It was noticed that the whole Connect keyword can fail if Login
of single iteration fails. Avoiding this.

If switch of ethernet interface from ghaf-host to net-vm
happens between Open Connection and Login keywords it is possible
that the opened connection is actually lost in between and thus
Login fails.

Signed-off-by: Samuli Leivo <samuli.leivo@unikie.com>