Change plot_dir to relative path
In case of using non-default output directory the path
to performance plots was set absolute. If report.html
is opened from the location it was originally created
it can show the links to performance plots but if the
result folder is moved to another location the links
to plots are broken in report.html.

This commit sets relative path to performance plots also
for non-default output directory as it is the case for
default output directory.

Signed-off-by: Samuli Leivo <samuli.leivo@unikie.com>