Save debug screenshots
Add fail_expected and debug_screenshot arguments
to keyword 'Locate image on screen'. So far there
is only one keyword 'Open app menu' where it is
expected to fail (although not failing a test if
succeeding). On the other hand we don't want to
save screenshots from 'Check if locked' wether
True or False.

If the pass_status of Locate image on screen is
contrary to fail_expected and debug_screenshot=True
(default value) the current screenshot file is saved
with timestamp so that further tests won't overwrite it.

Saved and timestamped screenshots can help in debugging
gui problems in tests.

Signed-off-by: Samuli Leivo <samuli.leivo@unikie.com>