Fix bug in marginal calculation
Occasional close-to-zero results have led to situations
where lower_marginal is higher than upper_marginal. With
such limits it is impossible to pass the test.

Prevent this happening by removing secondary pass/fail
criteria based on the deviation from the first measurement
of the baseline period.

Renamed a variable for clarity.

Signed-off-by: Samuli Leivo <samuli.leivo@unikie.com>