Small fixes
Save journalctl.txt instead of journalctl.log in
serial_keyword.Save log for jenkins controller to
save it.

If a performance test has not yet valid measurement results
in the measurement history (all are below self.low_limit)
let's just plot a single marginal self.low_limit, not something
depending on the invalid measured values.
(upper_marginal should actually be N/A before baseline but as
we are plotting the marginals without legend we can just plot
the marginal twice.)

Signed-off-by: Samuli Leivo <samuli.leivo@unikie.com>