Show correct string as build type
Currently build_type gets always value "packages" in
the runs which have image url defined or include
building the image. In the manual runs where url is
not defined and the device is booted from an existing
image build_type gets directly the device tag.

Modify initialization of self.build_type so that it gets
something more reasonable:
- builder architecture if job variable includes full target name
- "unknown" if job variable includes only device tag

Signed-off-by: Samuli Leivo <samuli.leivo@unikie.com>